feat: add-board-route
