+++ src/App.tsx
@@ -7,6 +7,7 @@ import { Signup } from './pages/Signup';
 import { Login } from './pages/Login';
 import { Main } from './pages/Main';
 import { Profile } from './pages/Profile';
+import Board from './components/main/Board';
 
 import { useAppSelector } from './core/hooks/redux';
 
@@ -24,6 +25,7 @@ const App = () => {
           <Route path="/signup" element={token ? <Navigate replace to="/home" /> : <Signup />} />
           <Route path="/main" element={token ? <Main /> : <Navigate replace to="/home" />} />
           <Route path="/home" element={<Home />} />
+          <Route path="/board" element={<Board />} />
           <Route path="*" element={<NotFound />} />
         </Routes>
       </main>
